requirements.txt
Django==3.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
flake8==3.8.3
psycopg2==2.8.6
PyJWT==1.7.1
celery==4.4.7
redis==3.5.3
environ==1.0
